Edit checklist_form_page.dart
